requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
chromadb
langchain_google_genai
youtube_transcript_api
pathlib